completed feature spring-boot-api

+++ MoneyTransferSystem/src/main/java/com/fidelity/mts/application/dto/TransferRequest.java
@@ -3,7 +3,7 @@ package com.fidelity.mts.application.dto;
 import java.math.BigDecimal;
 import java.util.UUID;
 
-public record TransferRequest(Integer fromAccountId,Integer toAccountId,
+public record TransferRequest(long fromAccountId, long toAccountId,
 		BigDecimal amount, UUID idempotencyKey) {
 
 }
